Chore: Rename SelectWithLabel.tsx
